Move membership related files into membership dir
