+++ index.js
@@ -40,9 +40,9 @@ var Dissemination = require('./lib/dissemination.js');
 var errors = require('./lib/errors.js');
 var getTChannelVersion = require('./lib/util.js').getTChannelVersion;
 var HashRing = require('./lib/ring');
-var Membership = require('./lib/membership.js');
-var MembershipIterator = require('./lib/membership-iterator.js');
-var MembershipUpdateRollup = require('./lib/membership-update-rollup.js');
+var Membership = require('./lib/membership/index.js');
+var MembershipIterator = require('./lib/membership/iterator.js');
+var MembershipUpdateRollup = require('./lib/membership/rollup.js');
 var nulls = require('./lib/nulls');
 var rawHead = require('./lib/request-proxy/util.js').rawHead;
 var RequestProxy = require('./lib/request-proxy/index.js');
