+++ lib/membership-set-listener.js
@@ -19,7 +19,7 @@
 // THE SOFTWARE.
 'use strict';
 
-var Member = require('./member.js');
+var Member = require('./membership/member.js');
 
 module.exports = function createMembershipSetListener(ringpop) {
     return function onMembershipSet(updates) {
