+++ lib/membership/index.js
@@ -23,8 +23,8 @@ var _ = require('underscore');
 var EventEmitter = require('events').EventEmitter;
 var farmhash = require('farmhash');
 var Member = require('./member.js');
-var MembershipUpdateRules = require('./membership-update-rules.js');
-var mergeMembershipChangesets = require('./membership-changeset-merge.js');
+var MembershipUpdateRules = require('./rules.js');
+var mergeMembershipChangesets = require('./merge.js');
 var util = require('util');
 var uuid = require('node-uuid');
