+++ lib/membership/member.js
