+++ lib/membership/merge.js
