+++ lib/swim/join-response-merge.js
@@ -19,7 +19,7 @@
 // THE SOFTWARE.
 'use strict';
 
-var mergeMembershipChangesets = require('../membership-changeset-merge.js');
+var mergeMembershipChangesets = require('../membership/merge.js');
 
 function hasSameChecksums(joinResponses) {
     var lastChecksum = -1;
