+++ test/membership-update-rollup-test.js
@@ -20,7 +20,7 @@
 'use strict';
 
 var after = require('after');
-var MembershipUpdateRollup = require('../lib/membership-update-rollup.js');
+var MembershipUpdateRollup = require('../lib/membership/rollup.js');
 var Ringpop = require('../index.js');
 var test = require('tape');
